Dialogs and dropdown for feed added

+++ app/src/main/java/org/listenbrainz/android/repository/feed/FeedRepositoryImpl.kt
@@ -1,10 +1,10 @@
 package org.listenbrainz.android.repository.feed
 
-import org.listenbrainz.android.model.FeedData
-import org.listenbrainz.android.model.FeedEventDeletionData
-import org.listenbrainz.android.model.FeedEventVisibilityData
 import org.listenbrainz.android.model.ResponseError
 import org.listenbrainz.android.model.SocialResponse
+import org.listenbrainz.android.model.feed.FeedData
+import org.listenbrainz.android.model.feed.FeedEventDeletionData
+import org.listenbrainz.android.model.feed.FeedEventVisibilityData
 import org.listenbrainz.android.service.FeedService
 import org.listenbrainz.android.util.Resource
 import org.listenbrainz.android.util.Utils
@@ -15,12 +15,14 @@ class FeedRepositoryImpl @Inject constructor(
 ) : FeedRepository {
     
     override suspend fun getFeedEvents(
-        username: String,
+        username: String?,
         maxTs: Int?,
         minTs: Int?,
         count: Int
     ) : Resource<FeedData> =
         runCatching {
+            if (username.isNullOrEmpty()) return@runCatching Resource.failure(error = ResponseError.AUTH_HEADER_NOT_FOUND)
+            
             val response = service.getFeedEvents(
                 username = username,
                 maxTs = maxTs,
@@ -38,12 +40,14 @@ class FeedRepositoryImpl @Inject constructor(
     
     
     override suspend fun getFeedFollowListens(
-        username: String,
+        username: String?,
         maxTs: Int?,
         minTs: Int?,
         count: Int
     ): Resource<FeedData> =
         runCatching {
+            if (username.isNullOrEmpty()) return@runCatching Resource.failure(error = ResponseError.AUTH_HEADER_NOT_FOUND)
+            
             val response = service.getFeedFollowListens(
                 username = username,
                 maxTs = maxTs,
@@ -61,12 +65,14 @@ class FeedRepositoryImpl @Inject constructor(
     
     
     override suspend fun getFeedSimilarListens(
-        username: String,
+        username: String?,
         maxTs: Int?,
         minTs: Int?,
         count: Int
     ): Resource<FeedData> =
         runCatching {
+            if (username.isNullOrEmpty()) return@runCatching Resource.failure(error = ResponseError.AUTH_HEADER_NOT_FOUND)
+            
             val response = service.getFeedSimilarListens(
                 username = username,
                 maxTs = maxTs,
@@ -83,8 +89,10 @@ class FeedRepositoryImpl @Inject constructor(
         }.getOrElse { Utils.logAndReturn(it) }
     
     
-    override suspend fun deleteEvent(username: String, data: FeedEventDeletionData) : Resource<SocialResponse> =
+    override suspend fun deleteEvent(username: String?, data: FeedEventDeletionData) : Resource<SocialResponse> =
         runCatching {
+            if (username.isNullOrEmpty()) return@runCatching Resource.failure(error = ResponseError.AUTH_HEADER_NOT_FOUND)
+            
             val response = service.deleteEvent(
                 username = username,
                 body = data
@@ -99,8 +107,10 @@ class FeedRepositoryImpl @Inject constructor(
         }.getOrElse { Utils.logAndReturn(it) }
     
     
-    override suspend fun hideEvent(username: String, data: FeedEventVisibilityData) : Resource<SocialResponse> =
+    override suspend fun hideEvent(username: String?, data: FeedEventVisibilityData) : Resource<SocialResponse> =
         runCatching {
+            if (username.isNullOrEmpty()) return@runCatching Resource.failure(error = ResponseError.AUTH_HEADER_NOT_FOUND)
+            
             val response = service.hideEvent(
                 username = username,
                 body = data
@@ -115,8 +125,10 @@ class FeedRepositoryImpl @Inject constructor(
         }.getOrElse { Utils.logAndReturn(it) }
     
     
-    override suspend fun unhideEvent(username: String, data: FeedEventVisibilityData) : Resource<SocialResponse> =
+    override suspend fun unhideEvent(username: String?, data: FeedEventVisibilityData) : Resource<SocialResponse> =
         runCatching {
+            if (username.isNullOrEmpty()) return@runCatching Resource.failure(error = ResponseError.AUTH_HEADER_NOT_FOUND)
+            
             val response = service.unhideEvent(
                 username = username,
                 body = data
